+++ src/init.jl
@@ -1,5 +1,3 @@
-using Requires
-
 """
     AstroFITS.CFITSIO_VERSION
 
@@ -11,23 +9,15 @@ const CFITSIO_VERSION = VersionNumber(CFITSIO.CFITSIO_MAJOR,
                                       CFITSIO.CFITSIO_MINOR,
                                       CFITSIO.CFITSIO_MICRO)
 
+if !isdefined(Base, :get_extension)
+    using Requires
+end
+
 function __init__()
-    # Extend DataFrames when this package is loaded.
-    @require DataFrames="a93c6f00-e57d-5684-b7b6-d8193f3e46c0" begin
-        """
-            nrow(hdu::FitsTableHDU)
-
-        Return the number of rows of the FITS Table extension in `hdu`.
-
-        """
-        DataFrames.nrow(hdu::FitsTableHDU) = hdu.nrows
-
-        """
-            ncol(hdu::FitsTableHDU)
-
-        Return the number of columns of the FITS Table extension in `hdu`.
-
-        """
-        DataFrames.ncol(hdu::FitsTableHDU) = hdu.ncols
+    @static if !isdefined(Base, :get_extension)
+        # Extend DataFrames when this package is loaded.
+        @require DataFrames="a93c6f00-e57d-5684-b7b6-d8193f3e46c0" begin
+            include("../ext/AstroFITSDataFramesExt.jl")
+        end
     end
 end